flux_core::num

Module usizeSpec_5d009f273939a2ae

External specifications in core::num for usize.

This is not a real Rust item; it exists only to document the external specification.

impl usize
{
    /// Core impl: https://github.com/rust-lang/rust/blob/0e95a0f4c677002a5d4ac5bc59d97885e6f51f71/library/core/src/num/uint_macros.rs#L2384-L2400
    #[no_panic]
    #[spec(fn(num : usize, rhs : usize) -> usize
    [clamp(num - rhs, 0, usize :: MAX)])] fn saturating_sub(self, rhs : usize)
    -> usize;
    /// Core impl: https://github.com/rust-lang/rust/blob/0e95a0f4c677002a5d4ac5bc59d97885e6f51f71/library/core/src/num/uint_macros.rs#L2340-L2356
    #[no_panic]
    #[spec(fn(num : usize, rhs : usize) -> usize
    [clamp(num + rhs, 0, usize :: MAX)])] fn saturating_add(self, rhs : usize)
    -> usize;
    /// Core impl: https://github.com/rust-lang/rust/blob/c6a955468b025dbe3d1de3e8f3e30496d1fb7f40/library/core/src/num/uint_macros.rs#L531-L545
    #[no_panic]
    #[spec(fn(num : usize, rhs : usize) -> Option < usize [num + rhs] >
    [num + rhs <= usize :: MAX])] fn checked_add(self, rhs : usize) -> Option
    < usize > ;
    /// Core impl: https://github.com/rust-lang/rust/blob/c6a955468b025dbe3d1de3e8f3e30496d1fb7f40/library/core/src/num/uint_macros.rs#L698-L709
    #[no_panic]
    #[spec(fn(num : usize, rhs : usize) -> Option < usize [num - rhs] >
    [num >= rhs])] fn checked_sub(self, rhs : usize) -> Option < usize > ;
    /// # Incomplete: result is bounded but not precisely determined by the input value.
    /// Core impl: https://github.com/rust-lang/rust/blob/c6a955468b025dbe3d1de3e8f3e30496d1fb7f40/library/core/src/num/uint_macros.rs#L84-L86
    #[no_panic] #[spec(fn(num : usize) -> u32 {v : v <= usize :: BITS})] fn
    count_ones(self) -> u32;
    /// # Incomplete: result is bounded but not precisely determined by the input value.
    /// Core impl: https://github.com/rust-lang/rust/blob/c6a955468b025dbe3d1de3e8f3e30496d1fb7f40/library/core/src/num/uint_macros.rs#L106-L108
    #[no_panic] #[spec(fn(num : usize) -> u32 {v : v <= usize :: BITS})] fn
    count_zeros(self) -> u32;
    /// # Incomplete: result is bounded but not precisely determined by the input value.
    /// Core impl: https://github.com/rust-lang/rust/blob/c6a955468b025dbe3d1de3e8f3e30496d1fb7f40/library/core/src/num/uint_macros.rs#L121-L123
    #[no_panic] #[spec(fn(num : usize) -> u32 {v : v <= usize :: BITS})] fn
    leading_zeros(self) -> u32;
    /// # Incomplete: result is bounded but not precisely determined by the input value.
    /// Core impl: https://github.com/rust-lang/rust/blob/c6a955468b025dbe3d1de3e8f3e30496d1fb7f40/library/core/src/num/uint_macros.rs#L137-L139
    #[no_panic] #[spec(fn(num : usize) -> u32 {v : v <= usize :: BITS})] fn
    trailing_zeros(self) -> u32;
    /// # Incomplete: result is bounded but not precisely determined by the input value.
    /// Core impl: https://github.com/rust-lang/rust/blob/c6a955468b025dbe3d1de3e8f3e30496d1fb7f40/library/core/src/num/uint_macros.rs#L153-L155
    #[no_panic] #[spec(fn(num : usize) -> u32 {v : v <= usize :: BITS})] fn
    leading_ones(self) -> u32;
    /// # Incomplete: result is bounded but not precisely determined by the input value.
    /// Core impl: https://github.com/rust-lang/rust/blob/c6a955468b025dbe3d1de3e8f3e30496d1fb7f40/library/core/src/num/uint_macros.rs#L169-L171
    #[no_panic] #[spec(fn(num : usize) -> u32 {v : v <= usize :: BITS})] fn
    trailing_ones(self) -> u32;
    /// Core impl: https://github.com/rust-lang/rust/blob/c6a955468b025dbe3d1de3e8f3e30496d1fb7f40/library/core/src/num/uint_macros.rs#L630-L633
    #[no_panic]
    #[spec(fn(num : usize, rhs : usize) -> Option < usize [num * rhs] >
    [num * rhs <= usize :: MAX])] fn checked_mul(self, rhs : usize) -> Option
    < usize > ;
    /// Core impl: https://github.com/rust-lang/rust/blob/c6a955468b025dbe3d1de3e8f3e30496d1fb7f40/library/core/src/num/uint_macros.rs#L720-L727
    #[no_panic]
    #[spec(fn(num : usize, rhs : usize) -> Option < usize [num / rhs] >
    [rhs > 0])] fn checked_div(self, rhs : usize) -> Option < usize > ;
    /// Core impl: https://github.com/rust-lang/rust/blob/c6a955468b025dbe3d1de3e8f3e30496d1fb7f40/library/core/src/num/uint_macros.rs#L2210-L2212
    #[no_panic]
    #[spec(fn(num : usize, rhs : usize) -> usize
    [wrap_once(num + rhs, 0, usize :: MAX)])] fn
    wrapping_add(self, rhs : usize) -> usize;
    /// Core impl: https://github.com/rust-lang/rust/blob/c6a955468b025dbe3d1de3e8f3e30496d1fb7f40/library/core/src/num/uint_macros.rs#L2260-L2262
    #[no_panic]
    #[spec(fn(num : usize, rhs : usize) -> usize
    [wrap_once(num - rhs, 0, usize :: MAX)])] fn
    wrapping_sub(self, rhs : usize) -> usize;
}

Modules

checked_add

Original declaration.

checked_div

Original declaration.

checked_mul

Original declaration.

checked_sub

Original declaration.

count_ones

Original declaration.

count_zeros

Original declaration.

leading_ones

Original declaration.

leading_zeros

Original declaration.

saturating_add

Original declaration.

saturating_sub

Original declaration.

trailing_ones

Original declaration.

trailing_zeros

Original declaration.

wrapping_add

Original declaration.

wrapping_sub

Original declaration.