flux_core::ptr::non_null

Module NonNullSpec_764f81474195cf53

External specifications in core::ptr for NonNull < T >.

This is not a real Rust item; it exists only to document the external specification.

impl<T> NonNull<T> {
    /// Need to enforce that the pointer value is non-null, because as the name suggests, this function does not check.
    /// Core impl: https://github.com/rust-lang/rust/blob/c871d09d1cc32a649f4c5177bb819646260ed120/library/core/src/ptr/non_null.rs#L234
    #[spec(fn(*mut[@p] T) -> NonNull<T>[p.base, p.addr, p.size] requires p.addr != 0)]
    unsafe fn new_unchecked(ptr: *mut T) -> Self;

    /// Core impl: https://github.com/rust-lang/rust/blob/c871d09d1cc32a649f4c5177bb819646260ed120/library/core/src/ptr/non_null.rs#L270
    #[spec(fn(*mut[@p] T) -> Option<NonNull<T>[p.base, p.addr, p.size]>[p.addr != 0])]
    fn new(ptr: *mut T) -> Option<NonNull<T>>;

    /// Core impl: https://github.com/rust-lang/rust/blob/c871d09d1cc32a649f4c5177bb819646260ed120/library/core/src/ptr/non_null.rs#L402
    #[spec(fn(NonNull<T>[@base, @addr, @size]) -> *mut[base, addr, size] T)]
    fn as_ptr(self) -> *mut T;

    /// Core impl: https://github.com/rust-lang/rust/blob/4b7e3a76d8df78960dc7c65cad43f5da1dac8ade/library/core/src/ptr/non_null.rs#L512
    #[spec(fn(NonNull<T>[@base, @addr, @size]) -> NonNull<U>[base, addr, size])]
    fn cast<U>(self) -> NonNull<U>;

    /// Core impl: https://github.com/rust-lang/rust/blob/c871d09d1cc32a649f4c5177bb819646260ed120/library/core/src/ptr/non_null.rs#L652
    #[spec(fn(NonNull<T>[@base, @addr, @size], count: usize)
        -> NonNull<T>[base, addr + count * T::size_of(), size - count * T::size_of()]
            requires nn_in_bounds(base, addr, size) && count * T::size_of() <= size)]
    unsafe fn add(self, count: usize) -> Self;

    /// Core impl: https://github.com/rust-lang/rust/blob/c871d09d1cc32a649f4c5177bb819646260ed120/library/core/src/ptr/non_null.rs#L729
    #[spec(fn(NonNull<T>[@base, @addr, @size], count: usize)
        -> NonNull<T>[base, addr - count * T::size_of(), size + count * T::size_of()]
            requires nn_in_bounds(base, addr, size) && count * T::size_of() <= addr - base && addr - count * T::size_of() != 0)]
    unsafe fn sub(self, count: usize) -> Self;

    /// Core impl: https://github.com/rust-lang/rust/blob/c871d09d1cc32a649f4c5177bb819646260ed120/library/core/src/ptr/non_null.rs#L576
    #[spec(fn(NonNull<T>[@base, @addr, @size], count: isize)
        -> NonNull<T>[base, addr + count * T::size_of(), size - count * T::size_of()]
            requires nn_in_bounds(base, addr, size) && count * T::size_of() <= size && addr + count * T::size_of() >= base && addr + count * T::size_of() != 0)]
    unsafe fn offset(self, count: isize) -> Self;

    /// Core impl: https://github.com/rust-lang/rust/blob/c871d09d1cc32a649f4c5177bb819646260ed120/library/core/src/ptr/non_null.rs#L678
    #[spec(fn(NonNull<T>[@base, @addr, @size], count: usize)
        -> NonNull<T>[base, addr + count, size - count]
            requires nn_in_bounds(base, addr, size) && count <= size)]
    unsafe fn byte_add(self, count: usize) -> Self;

    /// Core impl: https://github.com/rust-lang/rust/blob/c871d09d1cc32a649f4c5177bb819646260ed120/library/core/src/ptr/non_null.rs#L760
    #[spec(fn(NonNull<T>[@base, @addr, @size], count: usize)
        -> NonNull<T>[base, addr - count, size + count]
            requires nn_in_bounds(base, addr, size) && count <= addr - base && addr - count != 0)]
    unsafe fn byte_sub(self, count: usize) -> Self;

    /// Core impl: https://github.com/rust-lang/rust/blob/c871d09d1cc32a649f4c5177bb819646260ed120/library/core/src/ptr/non_null.rs#L602
    #[spec(fn(NonNull<T>[@base, @addr, @size], count: isize)
        -> NonNull<T>[base, addr + count, size - count]
            requires nn_in_bounds(base, addr, size) && count <= size && addr + count >= base && addr + count != 0)]
    unsafe fn byte_offset(self, count: isize) -> Self;

    /// Core impl: https://github.com/rust-lang/rust/blob/c871d09d1cc32a649f4c5177bb819646260ed120/library/core/src/ptr/non_null.rs#L858
    #[spec(fn(NonNull<T>[@sbase, @saddr, @ssize], origin: NonNull<T>[@obase, @oaddr, @osize]) -> isize[(saddr - oaddr) / T::size_of()]
        requires sbase == obase &&
                 nn_in_bounds(sbase, saddr, ssize) &&
                 nn_in_bounds(obase, oaddr, osize) &&
                 (saddr - oaddr) % T::size_of() == 0 &&
                 T::size_of() > 0)]
    unsafe fn offset_from(self, origin: NonNull<T>) -> isize
    where
        T: Sized;

    /// Core impl: https://github.com/rust-lang/rust/blob/c871d09d1cc32a649f4c5177bb819646260ed120/library/core/src/ptr/non_null.rs#L949
    #[spec(fn(NonNull<T>[@sbase, @saddr, @ssize], subtracted: NonNull<T>[@obase, @oaddr, @osize]) -> usize[(saddr - oaddr) / T::size_of()]
        requires sbase == obase &&
                 nn_in_bounds(sbase, saddr, ssize) &&
                 nn_in_bounds(obase, oaddr, osize) &&
                 saddr >= oaddr &&
                 (saddr - oaddr) % T::size_of() == 0 &&
                 T::size_of() > 0)]
    unsafe fn offset_from_unsigned(self, subtracted: NonNull<T>) -> usize
    where
        T: Sized;

    /// Core impl: https://github.com/rust-lang/rust/blob/c871d09d1cc32a649f4c5177bb819646260ed120/library/core/src/ptr/non_null.rs#L986
    #[spec(fn(NonNull<T>[@base, @addr, @size]) -> T
        requires nn_valid(base, addr, size, T::size_of()) && nn_aligned_to(addr, T::align_of()))]
    unsafe fn read(self) -> T
    where
        T: Sized;

    /// Core impl: https://github.com/rust-lang/rust/blob/c871d09d1cc32a649f4c5177bb819646260ed120/library/core/src/ptr/non_null.rs#L1141
    #[spec(fn(NonNull<T>[@base, @addr, @size], val: T)
        requires nn_valid(base, addr, size, T::size_of()) && nn_aligned_to(addr, T::align_of()))]
    unsafe fn write(self, val: T)
    where
        T: Sized;
}

Modules

add

Original declaration.

as_ptr

Original declaration.

byte_add

Original declaration.

byte_offset

Original declaration.

byte_sub

Original declaration.

cast

Original declaration.

new

Original declaration.

new_unchecked

Original declaration.

offset

Original declaration.

offset_from

Original declaration.

offset_from_unsigned

Original declaration.

read

Original declaration.

sub

Original declaration.

write

Original declaration.